The DMD r314 commit message is:
bugzilla 2029 (Typesafe variadic functions don't work in CTFE

The DMD r315 commit message is:
bugzilla 2029 - try again
---
 dmd/constfold.c   |   11 ++++-
 dmd/declaration.c |   21 +++++++++-
 dmd/declaration.h |   10 ++++-
 dmd/expression.c  |    1 +
 dmd/interpret.c   |  111 +++++++++++++++++++++++++++++++++++++++++++++--------
 dmd/mars.h        |    2 +-
 dmd/mtype.c       |    2 +-
 7 files changed, 135 insertions(+), 23 deletions(-)

+++ dmd/mars.h
@@ -417,7 +417,7 @@ enum MATCH
     MATCHexact		// exact match
 };
 
-typedef unsigned StorageClass;
+typedef uint64_t StorageClass;
 
 
 void warning(Loc loc, const char *format, ...) IS_PRINTF(2);
